+++ crates/cubecl-std/src/tensor/view/operations/slice.rs
@@ -10,7 +10,7 @@ impl<T: CubePrimitive, IO: SliceVisibility> ViewOperationsExpand<T, Coords1d>
     fn __expand_read_method(
         &self,
         scope: &mut Scope,
-        pos: ExpandElementTyped<u32>,
+        pos: ExpandElementTyped<usize>,
     ) -> <T>::ExpandType {
         <Self as ListExpand<T>>::__expand_read_method(self, scope, pos)
     }
@@ -18,7 +18,7 @@ impl<T: CubePrimitive, IO: SliceVisibility> ViewOperationsExpand<T, Coords1d>
     fn __expand_read_checked_method(
         &self,
         scope: &mut Scope,
-        pos: ExpandElementTyped<u32>,
+        pos: ExpandElementTyped<usize>,
     ) -> <T>::ExpandType {
         let len = self.__expand_len_method(scope);
         let in_bounds = lt::expand(scope, pos.clone(), len);
@@ -30,7 +30,7 @@ impl<T: CubePrimitive, IO: SliceVisibility> ViewOperationsExpand<T, Coords1d>
     fn __expand_read_masked_method(
         &self,
         scope: &mut Scope,
-        pos: ExpandElementTyped<u32>,
+        pos: ExpandElementTyped<usize>,
         mask_value: <T>::ExpandType,
     ) -> <T>::ExpandType {
         let len = self.__expand_len_method(scope);
@@ -42,7 +42,7 @@ impl<T: CubePrimitive, IO: SliceVisibility> ViewOperationsExpand<T, Coords1d>
     fn __expand_read_unchecked_method(
         &self,
         scope: &mut Scope,
-        pos: ExpandElementTyped<u32>,
+        pos: ExpandElementTyped<usize>,
     ) -> <T>::ExpandType {
         <Self as ListExpand<T>>::__expand_read_unchecked_method(self, scope, pos)
     }
@@ -50,25 +50,25 @@ impl<T: CubePrimitive, IO: SliceVisibility> ViewOperationsExpand<T, Coords1d>
     fn __expand_to_linear_slice_method(
         &self,
         scope: &mut Scope,
-        pos: ExpandElementTyped<u32>,
-        end: ExpandElementTyped<u32>,
+        pos: ExpandElementTyped<usize>,
+        end: ExpandElementTyped<usize>,
     ) -> SliceExpand<T, ReadOnly> {
         // Convert to exclusive end
-        let end = add::expand(scope, end, 1u32.into());
+        let end = add::expand(scope, end, 1usize.into());
         // Handling for shapes that are 0 in at least one dim, ensures the slice is not
         // negative length.
         let start = Min::__expand_min(scope, pos, end.clone());
         <Self as SliceOperatorExpand<T>>::__expand_slice_method(self, scope, start, end)
     }
 
-    fn __expand_shape_method(&self, scope: &mut Scope) -> ExpandElementTyped<u32> {
+    fn __expand_shape_method(&self, scope: &mut Scope) -> ExpandElementTyped<usize> {
         <Self as ListExpand<T>>::__expand_len_method(self, scope)
     }
 
     fn __expand_is_in_bounds_method(
         &self,
         scope: &mut Scope,
-        pos: ExpandElementTyped<u32>,
+        pos: ExpandElementTyped<usize>,
     ) -> ExpandElementTyped<bool> {
         let len = self.__expand_shape_method(scope);
         lt::expand(scope, pos, len)
@@ -79,7 +79,7 @@ impl<T: CubePrimitive, IO: SliceVisibility> ViewOperationsExpand<T, Coords1d>
         _scope: &mut Scope,
         _barrier: BarrierExpand,
         _shared_memory: SliceExpand<T, ReadWrite>,
-        _pos: ExpandElementTyped<u32>,
+        _pos: ExpandElementTyped<usize>,
     ) {
         unimplemented!("Not a tensor map");
     }
@@ -90,7 +90,7 @@ impl<T: CubePrimitive> ViewOperationsMutExpand<T, Coords1d> for SliceExpand<T, R
     fn __expand_write_method(
         &self,
         scope: &mut Scope,
-        pos: ExpandElementTyped<u32>,
+        pos: ExpandElementTyped<usize>,
         value: <T>::ExpandType,
     ) {
         <Self as ListMutExpand<T>>::__expand_write_method(self, scope, pos, value)
@@ -99,7 +99,7 @@ impl<T: CubePrimitive> ViewOperationsMutExpand<T, Coords1d> for SliceExpand<T, R
     fn __expand_write_checked_method(
         &self,
         scope: &mut Scope,
-        pos: ExpandElementTyped<u32>,
+        pos: ExpandElementTyped<usize>,
         value: <T>::ExpandType,
     ) {
         let len = <Self as ListExpand<T>>::__expand_len_method(self, scope);
@@ -112,11 +112,11 @@ impl<T: CubePrimitive> ViewOperationsMutExpand<T, Coords1d> for SliceExpand<T, R
     fn __expand_to_linear_slice_mut_method(
         &self,
         scope: &mut Scope,
-        pos: ExpandElementTyped<u32>,
-        end: ExpandElementTyped<u32>,
+        pos: ExpandElementTyped<usize>,
+        end: ExpandElementTyped<usize>,
     ) -> SliceExpand<T, ReadWrite> {
         // Convert to exclusive end
-        let end = add::expand(scope, end, 1u32.into());
+        let end = add::expand(scope, end, 1usize.into());
         // Handling for shapes that are 0 in at least one dim, ensures the slice is not
         // negative length.
         let start = Min::__expand_min(scope, pos, end.clone());
